MutateX/requirements.txt
# Requirements for enhanced MutateX position list generator
biopython>=1.79 